datamining/requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-genai
google-cloud-storage
google-cloud-firestore
google-cloud-logging
vertexai
python-dotenv
requests
